# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "Microservice Posts",
    "description": "Microservice from the \"If community\" project for managing posts and their categories and interactions (upvotes)"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "api-token",
      "description": "An API token is a unique access key that grants permission to interact with the API, ensuring security and control over the operations performed."
    }
  },
  "paths": {
...
    "/votes": {
      "post": {
        "tags": [
          "Votes (Upvotes)"
        ],
        "description": "adding interation(upvote) user to microservice",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "postId": {
                  "type": "number",
                  "example": 1
                },
                "upvote": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "postId": {
                  "type": "number",
                  "example": 1
                },
                "upvote": {
                  "type": "boolean",
                  "example": true
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "id": {
                  "type": "number",
                  "example": 1
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-20T00:20:37.690Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Votes (Upvotes)"
        ],
        "description": "search for a user by ID and a post ID",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "postId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 6
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-16T12:39:10.306Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                },
                "upvote": {
                  "type": "boolean",
                  "example": true
                },
                "userId": {
                  "type": "string",
                  "example": "1"
                },
                "postId": {
                  "type": "string",
                  "example": "1"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "voto não encontrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Votes (Upvotes)"
        ],
        "description": "adding interation(upvote) user to microservice",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "postId": {
                  "type": "number",
                  "example": 1
                },
                "upvote": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 6
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-16T12:39:10.306Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2024-05-16T12:40:27.328Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                },
                "upvote": {
                  "type": "boolean",
                  "example": false
                },
                "userId": {
                  "type": "string",
                  "example": "1"
                },
                "postId": {
                  "type": "string",
                  "example": "1"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Votes (Upvotes)"
        ],
        "description": "removes a vote from the specified user id and post id",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "postId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "delete": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "voto não encontrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}